Interfaces example.
It's loosing somewhere to put the added 'name' 'organisaton'
fields/getters. Something that's probably quite common to gecoders. They
aren't actually a part of the UPRN system though. So could be something
of their own, could be an interface for the OSPlaces or could just do as
all the other geocoders and add functions to the final class with no
interface.

Also not sure if they shouldn't be able to return NULL. Certainly the
OsGridReference as it has to return an object otherwise. The uprn could
return an empty string. Thought, does OS Places only return buildings,
not streets also (with usrn codes?)